Default to popup view + works through refresh, window closes on open
